Hack together a very basic text editor
N.B: This is not all working.

This adds a new ffi-backed component that represents a xi-core
document. It's a single document. This also adds a new view that
is backed by this document, and from which we paint lines.

The new view doesn't really accept user input, and it doesn't work
with the input handling stuff that already exists; as a next step
we can look at integrating them.